Move translations' namespace
