+++ src/Component/legacy/spec/Translation/Provider/ImmutableTranslationLocaleProviderSpec.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
 namespace spec\Sylius\Component\Resource\Translation\Provider;
 
 use PhpSpec\ObjectBehavior;
-use Sylius\Component\Resource\Translation\Provider\TranslationLocaleProviderInterface;
+use Sylius\Resource\Translation\Provider\TranslationLocaleProviderInterface;
 
 final class ImmutableTranslationLocaleProviderSpec extends ObjectBehavior
 {
