+++ src/Component/legacy/spec/Translation/TranslatableEntityLocaleAssignerSpec.php
@@ -14,9 +14,9 @@ declare(strict_types=1);
 namespace spec\Sylius\Component\Resource\Translation;
 
 use PhpSpec\ObjectBehavior;
-use Sylius\Component\Resource\Translation\Provider\TranslationLocaleProviderInterface;
-use Sylius\Component\Resource\Translation\TranslatableEntityLocaleAssignerInterface;
 use Sylius\Resource\Model\TranslatableInterface;
+use Sylius\Resource\Translation\Provider\TranslationLocaleProviderInterface;
+use Sylius\Resource\Translation\TranslatableEntityLocaleAssignerInterface;
 
 final class TranslatableEntityLocaleAssignerSpec extends ObjectBehavior
 {
