--- src/Component/legacy/src/Translation/Provider/TranslationLocaleProviderInterface.php
@@ -1,24 +1,0 @@
-<?php
-
-/*
- * This file is part of the Sylius package.
- *
- * (c) Sylius Sp. z o.o.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace Sylius\Component\Resource\Translation\Provider;
-
-interface TranslationLocaleProviderInterface
-{
-    /**
-     * @return string[]
-     */
-    public function getDefinedLocalesCodes(): array;
-
-    public function getDefaultLocaleCode(): string;
-}
