+++ src/Component/legacy/tests/Factory/TranslatableFactoryTest.php
@@ -18,10 +18,10 @@ use Prophecy\PhpUnit\ProphecyTrait;
 use Prophecy\Prophecy\ObjectProphecy;
 use Sylius\Component\Resource\Factory\TranslatableFactory;
 use Sylius\Component\Resource\Factory\TranslatableFactoryInterface as LegacyTranslatableFactoryInterface;
-use Sylius\Component\Resource\Translation\Provider\TranslationLocaleProviderInterface;
 use Sylius\Resource\Factory\FactoryInterface;
 use Sylius\Resource\Factory\TranslatableFactory as NewTranslatableFactory;
 use Sylius\Resource\Factory\TranslatableFactoryInterface;
+use Sylius\Resource\Translation\Provider\TranslationLocaleProviderInterface;
 
 final class TranslatableFactoryTest extends TestCase
 {
